[test] Improve frontend tests for image, list, radio, splitButton and tree property sections
