+++ packages/forms/frontend/sirius-components-forms/src/propertysections/__tests__/ImagePropertySection.test.tsx
@@ -13,36 +13,13 @@
 import { MockedProvider } from '@apollo/client/testing';
 import { cleanup, render } from '@testing-library/react';
 import { afterEach, expect, test, vi } from 'vitest';
-import { GQLImage } from '../../form/FormEventFragments.types';
 import { ImagePropertySection } from '../ImagePropertySection';
+import { imageWithMaxWidth, imageWithNoMaxWidth } from './ImagePropertySection.data';
 
-crypto.randomUUID = vi.fn(() => '48be95fc-3422-45d3-b1f9-d590e847e9e1');
-
-afterEach(() => cleanup());
-
-const imageWithMaxWidth: GQLImage = {
-  label: 'myImage',
-  url: 'https://www.eclipse.org/sirius/common_assets/images/logos/logo_sirius.png',
-  maxWidth: '42',
-  iconURL: [],
-  hasHelpText: false,
-  __typename: 'Image',
-  diagnostics: [],
-  id: 'imageId',
-  readOnly: false,
-};
-
-const imageWithNoMaxWidth: GQLImage = {
-  label: 'myImage',
-  url: 'https://www.eclipse.org/sirius/common_assets/images/logos/logo_sirius.png',
-  maxWidth: '',
-  iconURL: [],
-  hasHelpText: false,
-  __typename: 'Image',
-  diagnostics: [],
-  id: 'imageId',
-  readOnly: false,
-};
+afterEach(() => {
+  cleanup();
+  vi.clearAllMocks();
+});
 
 test('render image widget with maxWidth', () => {
   render(
