+++ packages/forms/frontend/sirius-components-forms/src/propertysections/__tests__/ListPropertySection.test.tsx
@@ -10,7 +10,7 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import { MockedProvider, MockedResponse } from '@apollo/client/testing';
+import { MockedProvider } from '@apollo/client/testing';
 import {
   MessageOptions,
   Selection,
@@ -18,90 +18,18 @@ import {
   ToastContext,
   ToastContextValue,
 } from '@eclipse-sirius/sirius-components-core';
-import { act, cleanup, render, screen, waitFor } from '@testing-library/react';
+import { cleanup, render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { afterEach, expect, test, vi } from 'vitest';
-import { GQLList, GQLListItem } from '../../form/FormEventFragments.types';
-import { ListPropertySection, clickListItemMutation } from '../ListPropertySection';
-import {
-  GQLClickListItemMutationData,
-  GQLClickListItemMutationVariables,
-  GQLSuccessPayload,
-} from '../ListPropertySection.types';
+import { ListPropertySection } from '../ListPropertySection';
+import { listReadOnly, list, itemClickCalledCalledSuccessMock } from './ListPropertySection.data';
 
 const setSelection = () => {};
 
-crypto.randomUUID = vi.fn(() => '48be95fc-3422-45d3-b1f9-d590e847e9e1');
-
-afterEach(() => cleanup());
-
-const defaultListItems: GQLListItem[] = [
-  {
-    id: 'item1',
-    label: 'item1Label',
-    kind: 'itemKind',
-    iconURL: [],
-    deletable: true,
-  },
-  {
-    id: 'item2',
-    label: 'item2Label',
-    kind: 'itemKind',
-    iconURL: [],
-    deletable: true,
-  },
-  {
-    id: 'item3',
-    label: 'item3Label',
-    kind: 'itemKind',
-    iconURL: [],
-    deletable: true,
-  },
-];
-const defaultList = {
-  label: 'myList',
-  iconURL: [],
-  hasHelpText: false,
-  items: defaultListItems,
-  style: null,
-  __typename: 'List',
-  diagnostics: [],
-  id: 'listId',
-  readOnly: false,
-};
-
-const defaultListWithStyle: GQLList = {
-  ...defaultList,
-  style: {
-    color: 'RebeccaPurple',
-    bold: true,
-    fontSize: 20,
-    italic: true,
-    strikeThrough: true,
-    underline: true,
-  },
-};
-
-const readOnlyList: GQLList = {
-  ...defaultListWithStyle,
-  readOnly: true,
-};
-
-const clickListItemVariables: GQLClickListItemMutationVariables = {
-  input: {
-    id: '48be95fc-3422-45d3-b1f9-d590e847e9e1',
-    editingContextId: 'editingContextId',
-    representationId: 'formId',
-    listId: 'listId',
-    listItemId: 'item3',
-    clickEventKind: 'SINGLE_CLICK',
-  },
-};
-
-const successPayload: GQLSuccessPayload = { messages: [], __typename: 'SuccessPayload' };
-const clickListItemSuccessData: GQLClickListItemMutationData = {
-  clickListItem: successPayload,
-};
+afterEach(() => {
+  cleanup();
+  vi.clearAllMocks();
+});
 
 const mockEnqueue = vi.fn<[string, MessageOptions?], void>();
 
@@ -114,28 +42,11 @@ const emptySelection: Selection = {
 };
 
 test('should the click event sent on item click', async () => {
-  let clickListItemMutationCalled = false;
-  const itemClickCalledCalledSuccessMock: MockedResponse<Record<string, any>> = {
-    request: {
-      query: clickListItemMutation,
-      variables: clickListItemVariables,
-    },
-    result: () => {
-      clickListItemMutationCalled = true;
-      return { data: clickListItemSuccessData };
-    },
-  };
-
   render(
     <MockedProvider mocks={[itemClickCalledCalledSuccessMock]}>
       <ToastContext.Provider value={toastContextMock}>
         <SelectionContext.Provider value={{ selection: emptySelection, setSelection, selectedRepresentations: [] }}>
-          <ListPropertySection
-            editingContextId="editingContextId"
-            formId="formId"
-            widget={defaultListWithStyle}
-            readOnly={false}
-          />
+          <ListPropertySection editingContextId="editingContextId" formId="formId" widget={list} readOnly={false} />
         </SelectionContext.Provider>
       </ToastContext.Provider>
     </MockedProvider>
@@ -143,29 +54,13 @@ test('should the click event sent on item click', async () => {
   const element: HTMLElement = screen.getByTestId('representation-item3Label');
   expect(element.textContent).toBe('item3Label');
 
-  await act(async () => {
-    userEvent.click(element);
-    await new Promise((resolve) => setTimeout(resolve, 0));
-
-    await waitFor(() => {
-      expect(clickListItemMutationCalled).toBeTruthy();
-    });
+  await userEvent.click(element);
+  await waitFor(() => {
+    expect(itemClickCalledCalledSuccessMock.result).toHaveBeenCalledTimes(1);
   });
 });
 
 test('when in read-only mode, list items can still be selected but handlers are not invoked', async () => {
-  let clickListItemMutationCalled = false;
-  const itemClickCalledCalledSuccessMock: MockedResponse<Record<string, any>> = {
-    request: {
-      query: clickListItemMutation,
-      variables: clickListItemVariables,
-    },
-    result: () => {
-      clickListItemMutationCalled = true;
-      return { data: clickListItemSuccessData };
-    },
-  };
-
   render(
     <MockedProvider mocks={[itemClickCalledCalledSuccessMock]}>
       <ToastContext.Provider value={toastContextMock}>
@@ -173,7 +68,7 @@ test('when in read-only mode, list items can still be selected but handlers are 
           <ListPropertySection
             editingContextId="editingContextId"
             formId="formId"
-            widget={readOnlyList}
+            widget={listReadOnly}
             readOnly={false}
           />
         </SelectionContext.Provider>
@@ -184,12 +79,8 @@ test('when in read-only mode, list items can still be selected but handlers are 
   const element: HTMLElement = screen.getByTestId('representation-item3Label');
   expect(element.textContent).toBe('item3Label');
 
-  await act(async () => {
-    userEvent.click(element);
-    await new Promise((resolve) => setTimeout(resolve, 0));
-
-    await waitFor(() => {
-      expect(clickListItemMutationCalled).toBeFalsy();
-    });
+  await userEvent.click(element);
+  await waitFor(() => {
+    expect(itemClickCalledCalledSuccessMock.result).toHaveBeenCalledTimes(0);
   });
 });
