+++ packages/forms/frontend/sirius-components-forms/src/propertysections/__tests__/TextfieldPropertySection.test.tsx
@@ -23,7 +23,10 @@ import {
   textfieldReadOnly,
 } from './TextfieldPropertySection.data';
 
-afterEach(() => cleanup());
+afterEach(() => {
+  cleanup();
+  vi.clearAllMocks();
+});
 
 const mockEnqueue = vi.fn<[string, MessageOptions?], void>();
